organizing file structure
organizing file structure; standardize source architecture
