--- sigma.io.dev/.gitignore
@@ -1,1 +1,0 @@
-/Debug/
