--- sigma.io.dev/main.c
@@ -1,13 +1,0 @@
-/*	Created on: Aug 16, 2024
- *	Author: Badkraft
- *	
- *	main.c: 
- */
-
-#include "../sigma.io/sigma_io.h"
-
-int main (int argc, char **argv) {
-	printf("Sigma.C IO Dev Environment\n");
-
-
-}
