--- sigma.io/sigma_io_directory.h
@@ -1,12 +1,0 @@
-/*	Created on: Aug 16, 2024
- *	Author: Badkraft
- *	
- *	sigma_io_directory.h: 
- */
-
-#ifndef _SIGMA_IO_DIRECTORY_H_
-#define _SIGMA_IO_DIRECTORY_H_
-
-
-
-#endif /* _SIGMA_IO_DIRECTORY_H_ */
