--- sigma.io/sigma_io_file.c
@@ -1,7 +1,0 @@
-/*	Created on: Aug 16, 2024
- *	Author: Badkraft
- *	
- *	sigma_io_file.c: 
- */
-
-
