+++ source/io/io_stream.c
@@ -1,0 +1,2 @@
+
+#include "../open/io_stream.h"
