+++ source/io/test.c
@@ -1,0 +1,8 @@
+#include <stdio.h>
+
+#include "../../testing/test.h"
+
+int main(int argc, char *argv[])
+{
+	print_header("OpenPlatform: IO\n");
+}
