+++ testing/test.h
@@ -1,0 +1,14 @@
+#ifndef _TESTING_H
+#define _TESTING_H
+
+#include <stdio.h>
+#include <assert.h>>
+
+#include "open/core.h"
+
+void write_header(char *header)
+{
+    printf("%s\n", header);
+}
+
+#endif //  _TESTING_H
